De-dupe FocusZone docs
