+++ apps/public-docsite-resources/src/components/pages/FocusZonePage.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import { DemoPage } from '../DemoPage';
 
-import { FocusZonePageProps } from '@fluentui/react-examples/lib/react/FocusZone/FocusZone.doc';
+import { FocusZonePageProps } from '@fluentui/react-examples/lib/react-focus/FocusZone/FocusZone.doc';
 
 export const FocusZonePage = (props: { isHeaderVisible: boolean }) => (
   <DemoPage
-    jsonDocs={require('@uifabric/api-docs/lib/pages/react/FocusZone.page.json')}
+    jsonDocs={require('@uifabric/api-docs/lib/pages/react-focus/FocusZone.page.json')}
     {...{ ...FocusZonePageProps, ...props }}
   />
 );
