+++ apps/public-docsite/src/pages/Controls/FocusZonePage/FocusZonePage.doc.ts
@@ -1,5 +1,5 @@
 import { TFabricPlatformPageProps } from '../../../interfaces/Platforms';
-import { FocusZonePageProps as ExternalProps } from '@fluentui/react-examples/lib/react/FocusZone/FocusZone.doc';
+import { FocusZonePageProps as ExternalProps } from '@fluentui/react-examples/lib/react-focus/FocusZone/FocusZone.doc';
 
 const related = require('!raw-loader!@fluentui/public-docsite/src/pages/Controls/FocusZonePage/docs/FocusZoneRelated.md') as string;
