+++ packages/react-examples/src/react-focus/FocusZone/FocusZone.Disabled.Example.tsx
@@ -2,8 +2,10 @@ import * as React from 'react';
 
 import { DefaultButton } from '@fluentui/react/lib/Button';
 import { FocusZone, FocusZoneDirection } from '@fluentui/react-focus';
-import { TextField } from '@fluentui/react/lib/TextField';
+import { TextField, ITextFieldStyles } from '@fluentui/react/lib/TextField';
 import { Stack } from '@fluentui/react/lib/Stack';
+
+const textFieldStyles: Partial<ITextFieldStyles> = { root: { width: 200 } };
 
 export const FocusZoneDisabledExample: React.FunctionComponent = () => {
   const tokens = { childrenGap: 20 };
@@ -14,11 +16,7 @@ export const FocusZoneDisabledExample: React.FunctionComponent = () => {
           <span>Enabled FocusZone: </span>
           <DefaultButton>Button 1</DefaultButton>
           <DefaultButton>Button 2</DefaultButton>
-          <TextField
-            placeholder="FocusZone TextField"
-            styles={{ root: { width: 200 } }}
-            ariaLabel="FocusZone TextField"
-          />
+          <TextField placeholder="FocusZone TextField" styles={textFieldStyles} ariaLabel="FocusZone TextField" />
           <DefaultButton>Button 3</DefaultButton>
         </Stack>
       </FocusZone>
