+++ packages/react-examples/src/react-focus/FocusZone/FocusZone.List.Example.tsx
@@ -4,7 +4,7 @@ import { useConst } from '@uifabric/react-hooks';
 import { TextField } from '@fluentui/react';
 import { Link } from '@fluentui/react/lib/Link';
 import { FocusZone, FocusZoneDirection } from '@fluentui/react-focus';
-import { DetailsRow, IColumn, Selection, SelectionMode } from '@fluentui/react/lib/DetailsList';
+import { DetailsRow, IColumn, Selection, SelectionMode, IDetailsRowStyles } from '@fluentui/react/lib/DetailsList';
 
 const ITEMS = createArray(10, index => ({
   key: index.toString(),
@@ -42,6 +42,8 @@ const COLUMNS: IColumn[] = [
   },
 ];
 
+const detailsRowStyles: Partial<IDetailsRowStyles> = { root: { display: 'block', width: '100%' } };
+
 export const FocusZoneListExample: React.FunctionComponent = () => {
   //  Initialize the selection when the component is first rendered (same instance will be reused)
   const selection = useConst(() => {
@@ -65,7 +67,7 @@ export const FocusZoneListExample: React.FunctionComponent = () => {
           columns={COLUMNS}
           selectionMode={SelectionMode.none}
           selection={selection}
-          styles={{ root: { display: 'block', width: '100%' } }}
+          styles={detailsRowStyles}
         />
       ))}
     </FocusZone>
