+++ packages/react-examples/src/react-focus/FocusZone/FocusZone.Tabbable.Example.tsx
@@ -1,9 +1,24 @@
 import * as React from 'react';
 import { DefaultButton } from '@fluentui/react/lib/Button';
 import { FocusZone, FocusZoneDirection, FocusZoneTabbableElements } from '@fluentui/react-focus';
-import { TextField } from '@fluentui/react/lib/TextField';
+import { TextField, ITextFieldStyles } from '@fluentui/react/lib/TextField';
 import { Stack } from '@fluentui/react/lib/Stack';
 
+const textFieldStyles: Partial<ITextFieldStyles> = { root: { width: 200 } };
+const menuProps = {
+  items: [
+    {
+      key: 'emailMessage',
+      text: 'Email message',
+      iconProps: { iconName: 'Mail' },
+    },
+    {
+      key: 'calendarEvent',
+      text: 'Calendar event',
+      iconProps: { iconName: 'Calendar' },
+    },
+  ],
+};
 export const FocusZoneTabbableExample: React.FunctionComponent = () => {
   const tokens = { childrenGap: 20 };
   return (
@@ -17,31 +32,14 @@ export const FocusZoneTabbableExample: React.FunctionComponent = () => {
           <span>Circular Tabbable FocusZone: </span>
           <DefaultButton>Button 1</DefaultButton>
           <DefaultButton>Button 2</DefaultButton>
-          <TextField
-            placeholder="FocusZone TextField"
-            styles={{ root: { width: 200 } }}
-            ariaLabel="FocusZone TextField"
-          />
+          <TextField placeholder="FocusZone TextField" styles={textFieldStyles} ariaLabel="FocusZone TextField" />
           <DefaultButton>Button 3</DefaultButton>
           <DefaultButton
             text="Create account"
             split={true}
             onClick={alertClicked}
             splitButtonAriaLabel="See 2 sample options"
-            menuProps={{
-              items: [
-                {
-                  key: 'emailMessage',
-                  text: 'Email message',
-                  iconProps: { iconName: 'Mail' },
-                },
-                {
-                  key: 'calendarEvent',
-                  text: 'Calendar event',
-                  iconProps: { iconName: 'Calendar' },
-                },
-              ],
-            }}
+            menuProps={menuProps}
           />
         </Stack>
       </FocusZone>
@@ -54,11 +52,7 @@ export const FocusZoneTabbableExample: React.FunctionComponent = () => {
           <span>Input Only FocusZone: </span>
           <DefaultButton>Button 1</DefaultButton>
           <DefaultButton>Button 2</DefaultButton>
-          <TextField
-            placeholder="FocusZone TextField"
-            styles={{ root: { width: 200 } }}
-            ariaLabel="FocusZone TextField"
-          />
+          <TextField placeholder="FocusZone TextField" styles={textFieldStyles} ariaLabel="FocusZone TextField" />
           <DefaultButton>Button 3</DefaultButton>
         </Stack>
       </FocusZone>
