--- packages/react-examples/src/react/FocusZone/docs/FocusZoneOverview.md
@@ -1,3 +1,0 @@
-FocusZones abstract arrow key navigation behaviors. Tabbable elements (buttons, anchors, and elements with data-is-focusable='true' attributes) are considered when pressing directional arrow keys and focus is moved appropriately. Tabbing to a zone sets focus only to the current "active" element, making it simple to use the tab key to transition from one zone to the next, rather than through every focusable element.
-
-Using a FocusZone is simple. Just wrap a bunch of content inside of a FocusZone, and arrows and tabbing will be handled for you! See examples below.
