+++ packages/react-internal/src/FocusZone.ts
@@ -1,1 +1,1 @@
-export * from './components/FocusZone/index';
+export * from '@fluentui/react-focus';
