--- packages/react-internal/src/components/FocusZone/index.ts
@@ -1,1 +1,0 @@
-export * from '@fluentui/react-focus';
